requirements.txt
pandas
requests
google-cloud-bigquery
pandas-gbq
pyarrow
